+++ src/components/checkbox/checkbox.tsx
@@ -51,6 +51,7 @@ export const Checkbox = ({
     const [currentChecked, setCurrentChecked, setOriginalEvent] =
         useCurrentChecked({
             checked,
+            value,
             defaultChecked,
             controlType: ControlType.Checkbox,
             onChange,
@@ -71,7 +72,18 @@ export const Checkbox = ({
 
     function handleInputChange(e: ChangeEvent<HTMLInputElement>) {
         setOriginalEvent(e);
-        setCurrentChecked(e.target.checked);
+        if (checked === undefined) {
+            //非受控
+            setCurrentChecked(e.target.checked);
+        } else {
+            //受控
+            onChange?.({
+                e,
+                value,
+                controlType: ControlType.Checkbox,
+                checked: e.target.checked,
+            });
+        }
     }
 
     return (